Support reporting raids with SMS
If a subscribed user sends an SMS starting with "report", the text after
"report" will be interpreted as an address and sent out as an alert to
all subscribers. This operates through the same code path as the alert
form on the site.

Right now, this only supports plain text addresses. We will be able to
integrate it with geocoding in the future to acquire coordinates, as
well as correct typos.