Issue-2 - Move to MySQL DB
[Problem]
Currenty the alert system only supports an in memory database, which won't work for production environment. We need
something more persistent.

[Solution]
Changed to using MySQL schema and connection settings. Create instructions for running the service locally using:
  1. a locally running service and DB
  2. a locally running service with a Docker hosted DB
  3. a Docker hosted service and Docker hosted DB

Also created a *very* rough debug page for manual testing and made some formatting changes as I came across them.

[Test]
Tested the different approaches except for the locally running service and locally running DB if I'm being honest. I
prefer to do rapid dev with option (2).

I also added a debug page for testing and when adding new subscribers through it, checked the database in the Docker
container to make sure that the fields were being populated. Note that a lot of the fields on the debug page are being
ignored and the page could use with some improvement. It'll be a good follow up task.