Issue-7: Auth0 integration
[Problem]
The service requires authentication.

[Solution]
Instead of adding authentication into the service, use Auth0. This means that apps can directly use Auth0 to
authenticate and communicate with the service.

The Auth0 integration was written following this tutorial with some minor adjustments to improve code readbility.
* https://auth0.com/docs/quickstart/webapp/scala/01-login

For future work, the tutorial was using scala play 2.5.x while the project is using 2.6.x. It is using some deprecated
classes and methods, which should be updated. Additionally, I was trying to inject the AuthenticatedActionSupport into
the HTML pages, but couldn't get it to work so in the spirit of moving fast, I added it as a method parameter, which
wasn't too bad.

[Test]
* spun service up locally using Docker DB and ngrok callbacks with Auth0